Add PR survival statistics plot
I am not actually sure that indicating the overlap between merged and
closed is actually useful/meaningful.

Unfortunately, the current statics are missing a `closedAt` field,
so that it is not possible to include a "closed" line in the plot.

The code is a bit hackish in places, but I guess the whole code here
is, so that is probably fine :).