Fix off by 1 error in Pony factor calculation and logarithmic plotting (#37)
* Fix off by 1 error in Pony factor calculation

numpy.searchsorted determines an index N for where the threshold should
be inserted. So we need to add 1 to this index as the Pony factor
constitutes the majority > 50% of all PRs.

* Ensure that contributor with most PRs is displayed

ax.plot(np.cumsum(num_merged_prs_per_author), ".") creates an automatic
index starting with 0. However, because the x axis uses the log scale,
the first value at 0 is never shown! To fix this we pass in our own
index starting with 1.